Authorization with middleware and error message

+++ app/Http/Controllers/Backend/ModuleController.php
@@ -6,9 +6,10 @@ use App\Models\Module;
 use Illuminate\Support\Str;
 use Illuminate\Http\Request;
 use App\Http\Controllers\Controller;
+use Brian2694\Toastr\Facades\Toastr;
+use Illuminate\Support\Facades\Gate;
 use App\Http\Requests\ModuleStoreRequest;
 use App\Http\Requests\ModuleUpdateRequest;
-use Brian2694\Toastr\Facades\Toastr;
 
 class ModuleController extends Controller
 {
@@ -19,6 +20,9 @@ class ModuleController extends Controller
      */
     public function index()
     {
+        //authorize this user to access/give access to admin dashboard
+        Gate::authorize('index-module');
+
         $modules = Module::latest('id')
         ->select(['id', 'module_name', 'module_slug', 'updated_at'])
         ->paginate(20);
@@ -33,6 +37,9 @@ class ModuleController extends Controller
      */
     public function create()
     {
+        //authorize this user to access/give access to admin dashboard
+        Gate::authorize('create-module');
+
         return view('admin.pages.module.create');
     }
 
@@ -44,6 +51,9 @@ class ModuleController extends Controller
      */
     public function store(ModuleStoreRequest $request)
     {
+        //authorize this user to access/give access to admin dashboard
+        Gate::authorize('create-module');
+
         Module::create([
             'module_name' => $request->module_name,
             'module_slug' => Str::slug($request->module_name),
@@ -72,6 +82,9 @@ class ModuleController extends Controller
      */
     public function edit($module_slug)
     {
+        //authorize this user to access/give access to admin dashboard
+        Gate::authorize('edit-module');
+
         $module = Module::where('module_slug', $module_slug)->first();
         return view('admin.pages.module.edit', compact('module'));
     }
@@ -85,6 +98,9 @@ class ModuleController extends Controller
      */
     public function update(ModuleUpdateRequest $request, $module_slug)
     {
+        //authorize this user to access/give access to admin dashboard
+        Gate::authorize('edit-module');
+
         $module = Module::where('module_slug', $module_slug)->first();
         $module->update([
             'module_name' => $request->module_name,
@@ -103,6 +119,9 @@ class ModuleController extends Controller
      */
     public function destroy($module_slug)
     {
+        //authorize this user to access/give access to admin dashboard
+        Gate::authorize('delete-module');
+
         $module = Module::where('module_slug', $module_slug)->first();
         $module->delete();
