+++ app/Http/Controllers/Backend/RoleController.php
@@ -8,6 +8,7 @@ use Illuminate\Support\Str;
 use Illuminate\Http\Request;
 use App\Http\Controllers\Controller;
 use Brian2694\Toastr\Facades\Toastr;
+use Illuminate\Support\Facades\Gate;
 use App\Http\Requests\RoleStoreRequest;
 use App\Http\Requests\RoleUpdateRequest;
 
@@ -20,6 +21,9 @@ class RoleController extends Controller
      */
     public function index()
     {
+        //authorize this user to access/give access to admin dashboard
+        Gate::authorize('index-role');
+
         $roles = Role::with(['permissions:id,permission_name,permission_slug'])
         ->select(['id', 'role_name', 'role_slug', 'role_note', 'is_deletable', 'updated_at'])
         ->paginate(20);
@@ -34,6 +38,9 @@ class RoleController extends Controller
      */
     public function create()
     {
+        //authorize this user to access/give access to admin dashboard
+        Gate::authorize('create-role');
+
         $modules = Module::with(['permissions:id,module_id,permission_name,permission_slug'])
         ->select('id', 'module_name')->get();
         return view('admin.pages.role.create', compact('modules'));
@@ -47,6 +54,9 @@ class RoleController extends Controller
      */
     public function store(RoleStoreRequest $request)
     {
+        //authorize this user to access/give access to admin dashboard
+        Gate::authorize('create-role');
+
         Role::updateOrCreate([
             'role_name' => $request->role_name,
             'role_slug' => Str::slug($request->role_name),
@@ -76,6 +86,9 @@ class RoleController extends Controller
      */
     public function edit($role_slug)
     {
+        //authorize this user to access/give access to admin dashboard
+        Gate::authorize('edit-role');
+
         $role = Role::where('role_slug', $role_slug)->first();
         $modules = Module::with(['permissions:id,module_id,permission_name,permission_slug'])
         ->select('id', 'module_name')->get();
@@ -92,6 +105,9 @@ class RoleController extends Controller
      */
     public function update(RoleUpdateRequest $request, $role_slug)
     {
+        //authorize this user to access/give access to admin dashboard
+        Gate::authorize('edit-role');
+
         // dd($request->all(), $role_slug);
         $role = Role::where('role_slug', $role_slug)->first();
         // $role->update([
@@ -115,6 +131,9 @@ class RoleController extends Controller
      */
     public function destroy($role_slug)
     {
+        //authorize this user to access/give access to admin dashboard
+        Gate::authorize('delete-role');
+
         $role = Role::where('role_slug', $role_slug)->first();
         if ($role->is_deletable){
             $role->delete();
